filter returned credentials when username is given; closes #4
See git-clone(1) for supported URL formats when cloning a repository.
Usernames are parsed by Git and passed to git-credential-keepassxc as a
separate field so there is no need for this project to parse the URLs.

git-credential-keepassxc will only try filtering by username when more
than 1 entries are returned by KeePassXC, and will discard the filtering
result if it contains no entries.