update configuration instructions
`command -v git` can return a path to a symlink, but
`git-credentials-keepassxc` will see the resolved path when deciding to
allow or deny the caller access. On systems where this is the case (such
as Arch), the suggested configuration will therefore not work. Resolving
the symlink with `readlink -f` fixes this."