Reintroduce read timeout
Previously to resolve [1], where the program hung when response from
KeePass/KeePassXC happens to be a multiple of receive buffer size, a
read timeout was introduced in [2] to break the read loop.

However this was later found to break credential retrieval when user
approval is required [3].

Now the timeout is reintroduced with one minor difference: it only
breaks the loop when it's already got some data, so that when it's
waiting for user approval and has not received any data, the loop keeps
running.

[1] https://github.com/Frederick888/git-credential-keepassxc/issues/35
[2] https://github.com/Frederick888/git-credential-keepassxc/commit/2d3355915e7140fcdbe0721b96e9a8842ff2f384
[3] https://github.com/Frederick888/git-credential-keepassxc/commit/21d6cab0f40955f6bbd5b74fea22b137ea627a8a